chore: use hatch to manage the project
This patch uses `hatch` to manage the project, rather than trying
to manually handle dependencies and environments, etc. Please see
the documentation for how to use `hatch` to work on the project.

This makes the minimum changes to CI to reduce disruptiveness. A
follow up patch will sync CI with development.