fix: remove cchardet as a dependency
`cchardet` hasn't had a commit in almost two years (8 days shy as
of this commit). [This issue](https://github.com/PyYoshi/cChardet/issues/81)
is a six month old bug showing that `cchardet` doesn't support
python 3.11, which is the most recent python.

`cchardet` is fast, certainly, though the usage here is not as
performance sensitive as would warrant its use over `chardet`,
e.g. a difference of 800ms is probably not something that is worth
keeping problematic dependencies around.