chore: simplify circleci config
CI config should be simple, with the least amount of indirection
as possible. It should also be easy to run the same commands
in development as it is in CI (and no, looking up
`.circleci/config.yml` to get the command doesn't count). With
the addition of `hatch`, this can all be greatly simplified into
a few small tasks.

A small but functional change is that linting is now performed
using `ruff`, which is much faster and much more thorough for
linting. `ruff` found lints that were currently being checked
by `pylama` and friends, so those lints were alse fixed.

Additionally, there is much in the way of infrastructure files
that are no longer needed after `hatch` was added and this patch
simplifies the circleci config, so they have been removed.

As a note: the tests leave a lot to be desired. However, with this
plumbing in place, tests and test infrastructure can be built out
and this config will just pick it up and run with it.