test: add initial attributes tests
This patch brings the total test coverage up to 30% as a result of
adding some testing for some of the functions in `attributes.py` -
its own coverage is now 37%.

Some of these functions will be difficult to test due to the
quasi-concurrent nature of attributes. This not an interface that
we should lean into at all, and is adding complexity (and, surely,
bugs) where we don't need it.