chore: save mypy and ruff caches
The `check` step of CI takes the longest, and it does so because
`ruff` and `mypy` both have to start from scratch in their analysis.
This patch keeps the caches from both commands so that they can be
reused in the next run. This sholud make the the `check` step
much quicker.